separate chapterListItem in reciter page
